+++ src/merging/merging_nnls.jl
@@ -520,8 +520,8 @@ function compute_lhs_and_rhs_rate_preserving!(nnls_merging, lhs_matrix,
             compute_cross_sections_only!(computed_cs, interaction, g, electron_neutral_interactions, neutral_species_index)
             lhs_matrix[nnls_merging.n_moments+1, col_index] = get_cs_elastic(electron_neutral_interactions, computed_cs, neutral_species_index) * g
             lhs_matrix[nnls_merging.n_moments+2, col_index] = get_cs_ionization(electron_neutral_interactions, computed_cs, neutral_species_index) * g
-            nnls_merging.rhs_vector[nnls_merging.n_moments+1] += get_cs_elastic(electron_neutral_interactions, computed_cs, neutral_species_index) * g * particles[i].w
-            nnls_merging.rhs_vector[nnls_merging.n_moments+2] += get_cs_ionization(electron_neutral_interactions, computed_cs, neutral_species_index) * g * particles[i].w
+            nnls_merging.rhs_vector[nnls_merging.n_moments+1] = nnls_merging.rhs_vector[nnls_merging.n_moments+1] + get_cs_elastic(electron_neutral_interactions, computed_cs, neutral_species_index) * g * particles[i].w
+            nnls_merging.rhs_vector[nnls_merging.n_moments+2] = nnls_merging.rhs_vector[nnls_merging.n_moments+2] + get_cs_ionization(electron_neutral_interactions, computed_cs, neutral_species_index) * g * particles[i].w
 
             col_index += 1
         end
@@ -915,6 +915,10 @@ is used instead and also multiplied by the variance multiplier.
 If the residual exceeds `max_err` or
 the number of non-zero elements in the solution vector is equal to the original number of particles,
 `-1` is returned to signify a failure of the merging algorithm.
+
+# References
+* G. Oblapenko, A Non-Negative Least Squares-based Approach for Moment-Preserving Particle Merging.
+    [arXiv preprint, 2025](https://doi.org/10.48550/arXiv.2412.12354).
 """
 function merge_nnls_based!(rng, nnls_merging, particles, pia, cell, species;
                            vref=1.0, scaling=:variance,
@@ -922,7 +926,7 @@ function merge_nnls_based!(rng, nnls_merging, particles, pia, cell, species;
 
     # create LHS matrix
     n_add = centered_at_mean ? 1 : 0
-    n_add += 8 * length(v_multipliers)
+    n_add = n_add + 8 * length(v_multipliers)
     @inbounds lhs_ncols = pia.indexer[cell, species].n_local + n_add + n_rand_pairs
     lhs_matrix = zeros(nnls_merging.n_moments, lhs_ncols)
     nnls_merging.vref = vref
@@ -1020,6 +1024,10 @@ the NNLS matrix and RHS corresponding to conservation of electron-neutral collis
 If the residual exceeds `max_err` or
 the number of non-zero elements in the solution vector is equal to the original number of particles,
 `-1` is returned to signify a failure of the merging algorithm.
+
+# References
+* G. Oblapenko, A Non-Negative Least Squares-based Approach for Moment-Preserving Particle Merging.
+    [arXiv preprint, 2025](https://doi.org/10.48550/arXiv.2412.12354).
 """
 function merge_nnls_based_rate_preserving!(rng, nnls_merging,
                                            interaction, electron_neutral_interactions, computed_cs,
@@ -1030,7 +1038,7 @@ function merge_nnls_based_rate_preserving!(rng, nnls_merging,
 
     # create LHS matrix
     n_add = centered_at_mean ? 1 : 0
-    n_add += 8 * length(v_multipliers)
+    n_add = n_add + 8 * length(v_multipliers)
     @inbounds lhs_ncols = pia.indexer[cell, species].n_local + n_add + n_rand_pairs
     lhs_matrix = zeros(nnls_merging.n_moments+2, lhs_ncols)
     nnls_merging.vref = vref