+++ src/merging/merging_roulette.jl
@@ -1,0 +1,70 @@
+"""
+    merge_roulette!(rng, particles, pia, cell, species, target_np)
+
+Perform roulette merging - delete random particles until target number of particles is reached, and re-weight remaining particles
+to conserve number density.
+
+# Positional arguments
+* `rng`: the random number generator instance
+* `particles`: the `ParticleVector` instance containing the particles to be merged
+* `pia`: the `ParticleIndexerArray` instance
+* `cell`: the index of the grid cell in which particles are being merged
+* `species`: the index of the species being merged
+* `target_np`: the target post-merge number of particles
+
+# References
+* J. Watrous, D.B. Seidel, C.H. Moore, W. McDoniel,
+    Improvements to Particle Merge Algorithms for Sandia National Laboratories Plasma Physics Modeling
+    Code, EMPIRE. [Presentation, 2023](https://www.osti.gov/servlets/purl/2431184).
+"""
+function merge_roulette!(rng, particles, pia, cell, species, target_np)
+    current_count = pia.indexer[cell, species].n_local
+
+    n_to_delete = current_count - target_np
+
+    w_deleted = 0.0
+
+    @inbounds for _ in 1:n_to_delete
+        i_delete = floor(Int64, rand(rng, Float64) * pia.indexer[cell, species].n_local)
+        i_delete = map_cont_index(pia.indexer[cell, species], i_delete)
+
+        w_deleted += particles[i_delete].w
+        delete_particle!(particles, pia, cell, species, i_delete)
+    end
+
+    # we most likely break continuity
+    # and checking that we only deleted from group2 in the last domain cell
+    # is kind of too much work for this merge
+    @inbounds pia.contiguous[species] = false
+
+    w_total = 0.0
+    @inbounds s1 = pia.indexer[cell,species].start1
+    @inbounds e1 = pia.indexer[cell,species].end1
+    @inbounds for i in s1:e1
+        w_total += particles[i].w
+    end
+
+    @inbounds if pia.indexer[cell, species].n_group2 > 0
+        @inbounds s2 = pia.indexer[cell,species].start2
+        @inbounds e2 = pia.indexer[cell,species].end2
+        @inbounds for i in s2:e2
+            w_total += particles[i].w
+        end
+    end
+
+    scale_factor = (w_total + w_deleted) / w_total
+
+    @inbounds s1 = pia.indexer[cell,species].start1
+    @inbounds e1 = pia.indexer[cell,species].end1
+    @inbounds for i in s1:e1
+        particles[i].w *= scale_factor
+    end
+
+    @inbounds if pia.indexer[cell, species].n_group2 > 0
+        @inbounds s2 = pia.indexer[cell,species].start2
+        @inbounds e2 = pia.indexer[cell,species].end2
+        @inbounds for i in s2:e2
+            particles[i].w *= scale_factor
+        end
+    end
+end